backend/requirements.txt
fastapi[standard]
SQLAlchemy[asyncio]==2.0.36
alembic==1.13.3
pydantic-settings==2.6.0
yarl==1.16.0
python-jose==3.3.0
sqlmodel==0.0.22
psycopg==3.2.3
psycopg2-binary==2.9.10
greenlet==3.1.1
aiosqlite==0.20.0
bcrypt==4.2.0

pytest==8.3.3
pytest_asyncio==0.24.0

# for worker
celery==5.4.0
redis==5.2.0
pytest-celery==1.1.3
numpy==1.26.4
pandas-ta==0.3.14b0
pandas==2.2.3
apimoex==1.4.0
yahoo-fin==0.8.9.1

black==24.10.0
flake8==7.1.1
nest_asyncio==1.6.0
